chore: move isTradingActive helper to swapper api helpers
